refactor: move files related to Host section
